galaxy-cli-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "galaxy-cli-agent"
version = "0.1.0"
description = "CLI-based agent for interacting with Galaxy using pydantic-ai"
readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [
    {name = "Galaxy Team"}
]
dependencies = [
    "pydantic>=2.0.0,<3.0.0",
    "pydantic-ai>=0.1.0",
    "httpx>=0.24.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
    "python-dotenv>=1.0.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.0.270"
]

[project.scripts]
galaxy-agent = "galaxy_cli_agent.cli:app"

[tool.ruff]
line-length = 100
target-version = "py39"
select = ["E", "F", "I"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true